+++ components/php/versions/8.3.33/internal_functions.c
@@ -30,6 +30,7 @@ extern zend_module_entry gpio_module_entry;
 extern zend_module_entry store_module_entry;
 extern zend_module_entry mem_module_entry;
 extern zend_module_entry sys_module_entry;
+extern zend_module_entry events_module_entry;
 #ifdef PHP_EXT_I2C_ENABLED
 extern zend_module_entry i2c_module_entry;
 #endif
@@ -127,6 +128,7 @@ static zend_module_entry * const php_builtin_extensions[] = {
 	&store_module_entry,
 	&mem_module_entry,
 	&sys_module_entry,
+	&events_module_entry,
 #ifdef PHP_EXT_I2C_ENABLED
 	&i2c_module_entry,
 #endif